scripts/server/prep-and-deploy-crowdsec-config/requirements.txt
gspread>=5.0.0
oauth2client>=4.0.0
google-api-python-client>=2.0.0
pyyaml>=6.0
requests>=2.25.0
urllib3>=1.26.0